Enhancement to cast or wrap
Cast or wrap generically adds an error message

'cast to eros.Error'

In the event a cast is necessary for Eros operations. This adds a
new optional set of messages that can be added to casts which would
otherwise get automagically made.

This set of strings will be concatenated, separated by commas ','
at the time of cast and become the Eros Error object's message.

Change log:
- Added variadic array of strings to CastOrWrap
- Added variadic array of strings to Check
- Added variadic array of strings to Result->Check